requirements.txt
click
requests
python-dotenv
rich
readchar
google-generativeai
